refactor: organize modules into subpackages
