+++ app.py
@@ -13,7 +13,7 @@ import streamlit as st
 # Import necessary components from the new src package
 from src.config import logger # Import logger early for potential use during imports
 try:
-    from src.ui_components import (
+from src.ui.ui_components import (
         render_sidebar,
         render_recording_section,
         render_upload_section,
@@ -67,7 +67,7 @@ def main():
         if use_agent_pipeline and not use_local_llm:
             # Ensure render_agent_settings is imported if it's not already automatically available
             # (it should be, as render_sidebar is from the same module)
-            from src.ui_components import render_agent_settings
+            from src.ui.ui_components import render_agent_settings
             agent_settings = render_agent_settings()
 
         # Create tabs for different functionalities
@@ -110,7 +110,7 @@ def main():
             st.write("This section allows for a side-by-side comparison of notes generated by the traditional pipeline and the new agent-based pipeline.")
 
             # Import the comparison view function
-            from src.ui_components import render_comparison_view
+            from src.ui.ui_components import render_comparison_view
 
             # Placeholder for transcript input - e.g., file uploader or select existing
             st.subheader("1. Provide Transcript for Comparison")
